requirements.txt
Zgd3k4 :- 6
a2ttREhQRG9rUIVabUVsdVF2SjRXdz09 :- 32
-------------------------------------------------------------------------------
7TcPWz :- 6
N1hnRGJHZlE4QXp6RjVXVGRnRGJQdz09 :- 32
-------------------------------------------------------------------------------
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.6
gunicorn==20.0.4
heroku==0.1.4
idna==2.10
mysql-connector==2.2.9
python-dateutil==1.5
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.3
